Move sorting and filtering enums to more specific locations based on use
