+++ client/utils/constants.js
@@ -1,22 +1,6 @@
-export const TABLE_SORT_ORDERS = {
-    ASC: 'ASCENDING',
-    DESC: 'DESCENDING'
-};
-
-export const DATA_MANAGEMENT_TABLE_SORT_OPTIONS = {
-    NAME: 'NAME',
-    ATS: 'ATS',
-    PHASE: 'PHASE'
-};
-
 export const TEST_PLAN_VERSION_PHASES = {
     RD: 'RD',
     DRAFT: 'DRAFT',
     CANDIDATE: 'CANDIDATE',
     RECOMMENDED: 'RECOMMENDED'
-};
-
-export const DATA_MANAGEMENT_TABLE_FILTER_OPTIONS = {
-    ALL: 'ALL',
-    ...TEST_PLAN_VERSION_PHASES
 };
